requirements.txt
asgiref==3.3.1
Django==3.1.4
django-taggit==1.3.0
djangorestframework==3.12.2
pytz==2020.4
sqlparse==0.4.1
